Updated spiral_ver_avg_pix2prof to be easier to add db fields.
The subfunction "setReturnVar" in spiral_verification_avg_pix2prof
can now create all the return variables either with their null values
or their initial fill/empty arrays.  This makes it easier to add
fields to db, since now they only need to be added to that subfunction
to be created in the correct order.

Also, created the subfunction cleanUpOutput which will automatically
clean up the five normal output variables (removing fill values)  plus
iterate through any fields in db and remove the same indices.

Lastly, corrected a bug in dealing with the MATLAB_DISPLAY variable -
since I immediately convert the string to a number, it would be a NaN
and not an empty string if the env. variable wasn't set. So I needed
to replace isempty with isnan to set it to 1 if the variable wasn't set.