Addition of profile category debugging GUI and related changes.
NEW FILES:
    tabulate_profile_categories - takes various output structures
from categorize_aerosol_profiles and turns them into two tables:
one lists the categorizations of the profiles, one lists the
criteria used for each test.

    aerosol_no2_prof_gui - a gui .fig and its corresponding .m
code file that allows the user to examine the categorizations
of all profiles in tables output from tabulate_profile_categories
and to directly plot any profile they want.

MODIFIED FILES:
    categorize_aerosol_profile - added some code to allow parsing
of optional inputs; this is in preparation to write a wrapping function
that will call this multiple times with different criteria and consider
the output of those runs before assigning a category to the profile.

    spiral_verification_avg_pix2prof - minor change, included some
debugging text to identify if a profile is using ground site data or
not.