Various bug fixes and updates to improve aerosol categorization. MODIFIED FILES:     categorize_aerosol_profile: Several changes:         1)  The boolean add2output is now initialized to TRUE outside             any if statements to ensure it defaults to TRUE.         2)  The case where the slope for the aerosol bin was 0 was             incorrectly referencing NO2 altitude and concentration             values due to a copy/paste mistake.  This has been             corrected.         3)  When testing which solution to the quadratic equation             falls in the bin known to contain the critical fraction             altitude, the test against the lower altitude bound             was changed to >= from > to make sure cases where the             critical altitude falls exactly on the border are             accounted for. This should not affect real profiles             which are unlikely to be that precise.         4)  Added a quality check that the top bin value < bottom             bin value for both profiles. This handles profiles that             had poor sampling above the boundary layer especially             in the Maryland DISCOVER-AQ campaign.         5)  "quadratic" was misspelled in some comments.
    multiple_categorize_profile: modifed to allow direct input of an
NO2 and aerosol profile along the same lines as categorize_aerosol_profile
already did.  Outputs a string of the categorization.  This is useful
for categorizing e.g. the Leitao 2010 model scenarios.

    select_campaign_ranges: changed to take the directory and dates from
merge_field_names given a specified campaign.

    bin_rolling_vertical_profile: Matlab syntax updates, "addParamValue"
methods changed to "addParameter" methods for the input parser.

    Run_Spiral_Verification: if set to fetch what profile numbers to run,
and the input is an empty matrix, this now passes essentially a fill value
(-127) as the profile number to spiral_verification_avg_pix2prof. This is
because spiral_verf. considers an empty matrix to mean run all profiles,
whereas when an empty matrix is passed to Run_Spiral it generally means
not to run any profiles.

    spiral_verification_avg_pix2prof: reversed the previous behaviour of
the ground site data quality flag. Before if the 12th bit of the quality
flag output was true, that meant that ground data is available for this
campaign, but was not applied. To simplify matters, this bit is now set
only if ground data is applied.

    plot_fit_line: changed to respect the existing limits of the plot,
rather than always plotting lines out to 10^17.