WIP on identifying biomass burning plumes. 2 new files:
NEW FILES:
    biomass_burning_scripts/calc_day_background - calculates the background
values for a given chemical species in OMI pressure bins, that is, for a day's
worth of data, it will calculate the background value for each OMI pressure bin.
(This is going to be renamed to calc_presbin_background shortly.)

    biomass_burning_scripts/find_campaign_fires - primary script that identifies
fires based on the enhancement of CO correlated with enhancement of HCN and ACN.
Testing incomplete, may not work; however it is being simplified shortly.