Plume identification works, for the given dummy merge data.
"NEW" FILES:
    biomass_burning_scripts/calc_presbin_background - the old
calc_day_background. Works exactly the same way, just renamed to make room for
the new calc_day_background.

    Workspaces/DummyFireMerge.mat - (binary) the constructed dataset I've been
using to test the plume ID for bugs. Not intended to guarentee that the method
will ID real plumes.

MODIFIED FILES:
    biomass_burning_scripts/calc_day_background - now actually does a true daily
background, ignoring altitude. Also calculates it slightly differently, it now
finds the 0.1 quantile and uses that as background. (This number can also be
overridden as an optional argument.) The standard deviation of the background is
now taken as that of all data less that the 0.1 quantile.

    biomass_burning_scripts/find_campaign_fire - several changes, many geared
towards using the true daily background instead of the binned one:
        1) Added variable to set the averaging period.
        2) All code that had been needed to handle the fact that backgrounds
were given by bin is gone. This includes averaging the std. deviations and major
changes to the calculate_enhancement subfunction.
        3) Now figures out ACN/HCN plumes by a straight 20 ppt difference to
background. This is completely off-the-cuff and will need to be tweaked when
looking at real data. The method of 3 std. dev. was rejected because it said
everything was a plume then.
        4) Bug fix around line 185 that used the wrong time for the end of the
block; this led to a total failure of the correlation test until it was fixed.
        5) Fixed it so that the output structure returns UTC values, not the
index of the averaging period.