Updates to utility functions calc_day_background and remove_merge_fills
MODIFIED FILES:
    biomass_burning_scripts/calc_day_background - now can accept the
species name as either one defined in merge_field_names or a field in
Merge.Data (will try merge_field_names first)

    utility_scripts/remove_merge_fields - now allows for the "Fill"
field to be the string "n/a" (case insensitive), which means that
asking it to get UTC does not throw an error. Also handles specially
if the field asked for is "LONGITUDE" (which needs to have all values
180 converted to the proper value).  This was already being done with
the fourth-output lon, but now will be done if longitude is asked for
explicitly (makes it easier if longitude is all we want, we don't have
to remember which number output it was). Consequently the code to
"fix" longitude values was put in a subfunction.