Some general improvements to find_campaign_fires
MODIFIED FILES:
    biomass_burning_scripts/find_campaign_fires - some bugs fixed,
some code added to ensure units are respected and to document the
criteria used. Specifically:
        1) minimum R^2 value needed for correlation now set as a
variable near the beginning and saved in the output structure.
        2) calculating the enhancement now outputs in requested mixing
ratio units. Depends on the convert_units.m file in my Utils repo.
        3) fixed error message displayed when fields were not found in
the Names structure - before it output the fields that were there
(instead of weren't, as it should and does now).