Aerosol profile extrapolation implemented.
MODIFIED FILES:
    satellite_verification_scripts/Run_Spiral_Verification - run time options
changed, no substatial changes to code.

    satellite_verification_scripts/spiral_verification_avg_pix2prof -
        1) A composite aerosol profile is now obtained for the day and the
median of the top 10 values is used to extrapolate to the tropopause. This
profile is then used to fill in the top of each individual aerosol profile as
needed.
        2) The extrapolation of the top 10 median was moved to the nested
function "extrapolate_top_10" to allow reuse of code. This enabled the section
handling the extrapolation of the NO2 and temperature profiles to be cleaned up
as well. Tests with DISCOVER-MD had no change in the BEHR vs aircraft scatter
plots, so this did not break anything.  A warning will be issued if the median
value being extrapolated to is greater than the last valid bin of the profile.
        3) Removed output of aer_prof_height, since that was a different attempt
to address the same problem (inconsistent profile heights).