Bug fixed in extrapolate_profile
MODIFIED FILES:
    utility_scripts/extrapolate_profile - now appropriately handles cases where
all data bins are NaN for the top extrapolation cases. Previously it would fail
to identify the last non-NaN bin and so not fill in any bins.  The "median"
method will error if all bins are NaN (a method could be implemented, but is not
worth the rewrite at this point - would need to rebin the original data and do
the extrapolation before removing bins below the surface). "fit" and both "wrf"
methods should work now.