Abortive attempt to manually tweak the uncertainties in the EMG fits.
MODIFIED FILES:
    Emissions/Plotting/calc_fit_param_uncert.m - renamed 'fit_frac_uncerts' to
'fit_frac_sds' since the uncertainties were the result of scaling the standard
deviations. Added a warning if the fitting uncertainties were unreasonably
small. Applied asymmetrical scaling factors to the uncertainties.

    Emissions/Plotting/compute_emg_emis_tau.m - require that the inputs A and X0
need to be row vectors and the uncertainties 2-by-n arrays. Propagates
asymmetric uncertainties by assuming each side can be propagated by the normal
rules.

    Emissions/Plotting/fit_var_plotting.m - fits subfunction modified to handle
asymmetric uncertainties. plot_fit_envelopes subfunction modified to produce
plots that were meant for publication. The subfunction that tabulated
uncertainties was modified to pass percent SD rather than percent CI because I
had directly scaled the SDs not the CIs.

    Emissions/emg_fit_montecarlo.m - corrected the lower bound of a, also found
a way to get it to work for the probability function exp(-ssresid) by centering
the function on the optimum solution, i.e. exp{-(ssresid - ssresid_opt)}. This
may be self-fulfilling though, but if it did find a value < optimum, the
probability would increase, so why not?

NEW FILES:
    Emissions/emg_sd_scaling_factors.m - just returns the 2x5 array of scaling
factors for the unexvar SDs.