Some more cleanup of new functions that never got comitted
NEW FILES:
    aerosol_distribution/sum_aitken_mode.m - this was to add up Aitken mode
aerosols from field campaigns in order to give me a way to judge if the
conversion to Aitken mode number from GEOS-Chem output was good enough to use
for WRF-Chem boundary conditions.

    satellite_verification_scripts/Run_Avg_Spiral_Verification.m - compares
time-averaged aircraft and satellite columns (using both the CVM and PSM methods
for the satellite columns) to see if the PSM method provides any advantage.