Few additional tweaks to allow finer control over the validation
MODIFIED FILES:
     AMF_tools/hourly_amf_sensitivity.m - replaced old (and deleted) "clip"
function with newer "clipmat".

    satellite_verification_scritps/Run_Spiral_Verification.m - added DEBUG_LEVEL
as an input parameter.

    satellite_vericiation_scripts/spiral_verification_avg_pix2prof.m - added
'usecomposite' parameter to force the comparison to use the composite profile
above the top of each profile, rather than extrapolating. Also made sure that
the unit conversion for the top of the profile was right even if using a ppb
profile.