Made spiral_verification unit-aware for NO2 concentrations.
As a side effect of this, a duplicate function for reading Merge
fields has been eliminated.

MODIFIED FILES:
    read_icart_files/read_merge_field.m - should not be used, will throw an
error if called; REMOVE_MERGE_FILLS() should be used instead.

    aerosol_distributions/sum_aitken_mode.m - call to READ_MERGE_FIELD()
replaced with REMOVE_MERGE_FILLS().

    sat_ver_scripts/spiral_verification_avg_pix2prof.m - now uses the unit-aware
feature of REMOVE_MERGE_FILLS() to convert NO2 mixing ratio into unscaled
parts-per-part. Only if the unit in the Merge structure cannot be converted to
ppp will the 'conv_fact' parameter be used. 'conv_fact' now has NaN as a default
value so that the user MUST pass a conversion factor consciously, and it won't
just assume pptv if the units in the Merge file are unknown. All constant values
used in relation to NO2 concentrations have been converted to unscaled mixing
ratio values.

NEW FILE:
    One-off scripts/fix_soas_no2_units.m - sets the NO2 units in the SOAS merge
files to 'ppbv', as they should be.